refactor(api): migrate local.guard to discord.guard
Rename local.guard.ts to discord.guard.ts and update implementation for Discord-based authentication.

Update guards/index.ts to export from discord.guard instead of local.guard.
