+++ apps/api/src/modules/auth/guards/index.ts
@@ -1,1 +1,1 @@
-export * from './local.guard';
+export * from './discord.guard';
