--- apps/api/src/modules/auth/guards/local.guard.ts
@@ -1,20 +1,0 @@
-import { Injectable, ExecutionContext, CanActivate } from '@nestjs/common';
-import { AuthGuard } from '@nestjs/passport';
-
-@Injectable()
-export class LocalAuthGuard extends AuthGuard('local') {
-  async canActivate(context: ExecutionContext): Promise<boolean> {
-    const activate = (await super.canActivate(context)) as boolean;
-    const request = context.switchToHttp().getRequest();
-    await super.logIn(request);
-    return activate;
-  }
-}
-
-@Injectable()
-export class LocalAuthenticatedGuard implements CanActivate {
-  async canActivate(context: ExecutionContext) {
-    const request = context.switchToHttp().getRequest();
-    return request.isAuthenticated();
-  }
-}
